Update CommentedConfiguration.java
Eliminated destructive errors when loading a yaml with errors, constructor will now return null if something bad happened.
Methods like load() and save() now returns a boolean. Makes it easier to tell if something bad has happened.
Added the ability to add comments, keys and values without being deleted on the next reboot.
Small step towards the generality of the values linked to a key, get () will no longer return a string but an object and everything will be saved as an object. This does not mean that Locations or any kind of Spigot are now supported, but it is the basis to start from.